feat(pocket-ui-wgpu): shared Blit compositor for upscales and overlays
